Renamed surfaceSegmentation to SurfaceSegmentation (ROS style) Split DataCoordinator and SurfaceSegmentation into header and source files.
