+++ godel_surface_detection/CMakeLists.txt
@@ -52,7 +52,9 @@ include_directories(
 ## Declare a cpp library
 add_library(${PROJECT_NAME} 
   src/detection/surface_detection.cpp
-  src/detection/surface_segmentation_utils.cpp
+  src/segmentation/surface_segmentation_utils.cpp
+  src/segmentation/surface_segmentation.cpp
+  src/coordination/data_coordinator.cpp
   src/scan/robot_scan.cpp
   src/interactive/interactive_surface_server.cpp
   src/scan/profilimeter_scan.cpp
